make std an optional dependency
