+++ Cargo.toml
@@ -7,3 +7,7 @@ authors = ["Hayley Deckers <hayley@fusion.engineering, im@fckn.gay>"]
 
 [dependencies]
 arrayvec = { version = "0.7.4", default-features = false }
+
+[features]
+default = []
+std = []
